requirements.txt
aiobotocore==3.1.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioitertools==0.13.0
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.31.0
attrs==25.4.0
authx==1.5.0
bcrypt==5.0.0
boto3==1.42.58
botocore==1.42.58
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
databases==0.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.7
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
frozenlist==1.8.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
minio==7.2.20
multidict==6.7.1
oauth2==1.9.0.post1
packaging==26.0
pluggy==1.6.0
propcache==0.4.1
py-partiql-parser==0.6.3
pyasn1==0.6.2
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.3.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
redis==7.1.1
requests==2.32.5
responses==0.26.0
rich==14.3.2
rich-toolkit==0.19.0
rignore==0.7.6
rsa==4.9.1
s3transfer==0.16.0
sentry-sdk==2.52.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.6
wrapt==2.1.1
xmltodict==1.0.4
yarl==1.22.0
